Python/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.txt
#
appdirs==1.4.3
atomicwrites==1.1.5
attrs==18.1.0
autopep8==1.3.5
backcall==0.1.0
black==18.6b2
certifi==2018.4.16
chardet==3.0.4
click==6.7
cycler==0.10.0
decorator==4.3.0
docopt==0.6.2
genson==1.0.1
idna==2.7
ipython-genutils==0.2.0
ipython==7.0.1
jedi==0.13.1
kiwisolver==1.0.1
matplotlib==2.2.3
more-itertools==4.2.0
numpy==1.15.2
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
pipreqs==0.4.9
pluggy==0.6.0
prompt-toolkit==2.0.6
ptyprocess==0.6.0
py-cpuinfo==4.0.0
py==1.5.3
pycodestyle==2.4.0
pygal==2.4.0
pygaljs==1.0.1
pygments==2.2.0
pymongo==3.7.2
pyparsing==2.2.0
pytest-benchmark==3.1.1
pytest==3.6.1
python-dateutil==2.7.2
pytz==2018.5
red-black-tree-mod==1.20
requests==2.19.1
schema==0.6.7
simplegeneric==0.8.1
six==1.11.0
toml==0.9.4
traitlets==4.3.2
urllib3==1.23
wcwidth==0.1.7
yapf==0.22.0
yarg==0.1.9
